+++ apps/barnepensjon-ui/src/components/application/about-you/AboutYou.tsx
@@ -14,7 +14,7 @@ import LoggedInUserInfo from './LoggedInUserInfo'
 import FormElement from '../../common/FormElement'
 import { IAboutYou } from '../../../types/person'
 import PaymentDetails from '../../common/PaymentDetails'
-import { ApplicantRole } from '../scenario/ScenarioSelection'
+import { ApplicantRole } from '../../FrontPage'
 import { StandardBreddeRHFSelect } from '../../common/rhf/RHFSelect'
 import useCountries from '../../../hooks/useCountries'
 import { Bredde } from '../../../utils/bredde'
