+++ apps/barnepensjon-ui/src/components/application/summary/Summary.tsx
@@ -9,7 +9,7 @@ import FormGroup from '../../common/FormGroup'
 import Navigation, { NavRow } from '../../common/Navigation'
 import StepHeading from '../../common/StepHeading'
 import { StepProps } from '../Dialogue'
-import { ApplicantRole, ApplicantSituation } from '../scenario/ScenarioSelection'
+import { ApplicantRole, ApplicantSituation } from '../../FrontPage'
 import { SummaryAboutChildren } from './fragments/SummaryAboutChildren'
 import { SummaryAboutDeceasedParent } from './fragments/SummaryAboutDeceasedParent'
 import { SummaryAboutLivingParent } from './fragments/SummaryAboutLivingParent'
