+++ apps/barnepensjon-ui/src/components/application/summary/fragments/SummaryAboutChildren.tsx
@@ -6,7 +6,7 @@ import { JaNeiVetIkke } from '../../../../api/dto/FellesOpplysninger'
 import useTranslation from '../../../../hooks/useTranslation'
 import { IAboutChildren, IChild, ParentRelationType } from '../../../../types/person'
 import { StepLabelKey, StepPath } from '../../../../utils/steps'
-import { ApplicantRole, ApplicantSituation } from '../../scenario/ScenarioSelection'
+import { ApplicantRole, ApplicantSituation } from '../../../FrontPage'
 import { AccordionItem } from '../AccordionItem'
 import { TextGroup, TextGroupJaNeiVetIkke } from '../TextGroup'
 import PaymentDetailsSummary from './PaymentDetailsSummary'
